+++ app/Models/CategoryRoom.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\HasMany;
+
+class CategoryRoom extends Model
+{
+    protected $fillable = [
+        'name'
+    ];
+
+    protected $hidden = [
+        'created_at',
+        'updated_at'
+    ];
+
+    public function chatRoom(): HasMany
+    {
+        return $this->hasMany(chatRoom::class);
+    }
+}